chore: upgrade ruff quality gates for SOLID enforcement
New lint rules added to pyproject.toml:
- C90: mccabe complexity (SRP - max 10)
- PLR: pylint refactor (too many args/returns/branches)
- ANN: type annotations (DI/LSP contracts)
- S: flake8-bandit (security, critical for service manager)
- ASYNC: async correctness
- PERF: performance anti-patterns
- FURB: modern Python idioms
- N/ARG/RET/PT/PIE/RSE/T20/FLY: code quality

Also:
- Target Python 3.12+ (was 3.11)
- requires-python bumped to >=3.12
- Per-file-ignores for test patterns (S101, ANN, ARG, etc.)
- Fixed easy lint issues (RUF046, RET504, S108 noqa, S314 noqa)
- 21 remaining errors are genuine tech debt (complexity, async)

All 1171 tests passing.

+++ src/gpumod/fetchers/gguf.py
@@ -221,7 +221,7 @@ class GGUFFetcher:
         if file_size_bytes == 0:
             return 0
         size_mb = file_size_bytes / (1024 * 1024)
-        return int(math.ceil(size_mb * _VRAM_OVERHEAD_FACTOR))
+        return math.ceil(size_mb * _VRAM_OVERHEAD_FACTOR)
 
     @staticmethod
     def _validate_path(file_path: str) -> None:
